EID-602 Split common test utilities into separate class.
